--- config.json
@@ -1,37 +1,0 @@
-{
-    "date": [
-        1
-    ],
-    "sess": [
-        1,
-        2
-    ],
-    "price": [
-        1,
-        2
-    ],
-    "real_name": [
-        1
-    ],
-    "nick_name": "",
-    "ticket_num": 1,
-    "viewer_person": [
-        1
-    ],
-    "driver_path": "C:\\Users\\admin\\Desktop\\damaihelper-main\\chromedriver.exe",
-    "damai_url": "https://www.damai.cn/",
-    "target_url": "https://m.damai.cn/damai/detail/item.html?itemId=门票ID",
-    "comment": {
-        "title": "comment 下的所有内容为自定义注释,无实际含义",
-        "date": "1",
-        "sess": "场次序号,优先选中的场次序号放在前,填写的场次序号若大于实际场次序号,则会选中实际场次序号最大的",
-        "price": "票档序号,优先选中的票档序号放在前,填写的票档序号若大于实际票档序号,则会选中实际票档序号最大的",
-        "real_name": "实名者序号,已经弃用",
-        "nick_name": "用户昵称,已经弃用",
-        "ticket_num": "购买票数,购买票数与观影人序号的数量务必一致",
-        "viewer_person": "观影人序号(预先添加实名观影人),优先选中的序号放在前,填写的序号若大于实际序号,则会放弃选中",
-        "driver_path": "驱动地址",
-        "damai_url": "大麦首页地址,用于登录",
-        "target_url": "购票的实际地址,需要使用手机端的地址,域名: https://m.damai.cn/ 开头"
-    }
-}
